refactor: move category and task packages under modules
- Update import paths for category and task packages to use the new
  "modules" directory structure for better modularization and clarity.
- Adjust formatting and alignment in struct tags for consistency.
- Fix minor spacing issues in query pagination fields.
- Update route and handler imports to reflect the new module paths.

This restructuring improves project organization and prepares the codebase
for easier scalability and maintenance.
